Address comments: move out of main.rs; fix CLI help text
